requirements.txt
fastapi
uvicorn
gitpython
faiss-cpu
openai
tiktoken
numpy
groq